refactor: erstatt path-aliaser med enkelt @src/*-alias
- Fjerner @features/*, @shared/*, @utils/* fra tsconfig.json
- Legger til ett samlet alias: @src/* → src/*
- Oppdaterer alle 50+ filer med relative .. imports til @src/-imports
- Inkluderer biome auto-fiks på berørte filer
