+++ src/features/aktuelt/aktueltTypes.ts
@@ -1,4 +1,4 @@
-import { Microfrontend } from "../../microfrontends/microfrontendTypes";
+import type { Microfrontend } from "@src/microfrontends/microfrontendTypes";
 
 export interface Aktuelt {
   offerStepup: boolean;
