+++ src/features/alert-island/AlertIsland.astro
@@ -12,7 +12,7 @@ import Varsler from "./varsler/Varsler.astro";
 </div>
 
 <script>
-  import { logEvent } from "../../utils/client/umami";
+  import { logEvent } from "@src/shared/utils/client/umami";
 
   const utkast = document.querySelector("#utkast");
   const varsler = document.querySelector("#varlser");
