+++ src/features/din-oversikt/DinOversikt.astro
@@ -1,14 +1,14 @@
 ---
 import { DIN_OVERSIKT_URL } from "astro:env/server";
 import { BodyShort } from "@navikt/ds-react";
-import ClientError from "@shared/client-error/ClientError";
-import { MultiStatusError } from "@utils/server/error";
-import { fetchData } from "@utils/server/fetch";
-import logger from "@utils/server/logger";
-import { getAudience, getOboToken } from "@utils/server/token";
-import MicrofrontendSSR from "../../microfrontends/Microfrontend.astro";
-import MicrofrontendSSRSkeleton from "../../microfrontends/MicrofrontendSkeleton.astro";
-import type { Language } from "../../utils/server/language";
+import MicrofrontendSSR from "@src/microfrontends/Microfrontend.astro";
+import MicrofrontendSSRSkeleton from "@src/microfrontends/MicrofrontendSkeleton.astro";
+import ClientError from "@src/shared/client-error/ClientError";
+import { MultiStatusError } from "@src/shared/utils/server/error";
+import { fetchData } from "@src/shared/utils/server/fetch";
+import type { Language } from "@src/shared/utils/server/language";
+import logger from "@src/shared/utils/server/logger";
+import { getAudience, getOboToken } from "@src/shared/utils/server/token";
 import styles from "./DinOversikt.module.css";
 import type { PersonalizedContent } from "./DinOversiktTypes";
 import { getProduktkortList } from "./dinOversiktUtils.ts";
