+++ src/features/din-oversikt/DinOversiktTypes.ts
@@ -1,4 +1,4 @@
-import { Microfrontend } from "../../microfrontends/microfrontendTypes";
+import type { Microfrontend } from "@src/microfrontends/microfrontendTypes";
 
 export interface PersonalizedContent {
   microfrontends: Microfrontend[];
