+++ src/features/din-oversikt/meldekort/Meldekort.astro
@@ -1,9 +1,9 @@
 ---
-import { getAudience, getOboToken } from "@utils/server/token";
-import { fetchHtml } from "@utils/server/fetch";
 import { MELDEKORT_URL } from "astro:env/server";
-import ClientError from "@shared/client-error/ClientError";
-import logger from "@utils/server/logger";
+import ClientError from "@src/shared/client-error/ClientError";
+import { fetchHtml } from "@src/shared/utils/server/fetch";
+import logger from "@src/shared/utils/server/logger";
+import { getAudience, getOboToken } from "@src/shared/utils/server/token";
 
 interface Props {
   language: string;
