+++ src/features/dokumenter/dokument/Dokument.astro
@@ -1,10 +1,10 @@
 ---
 import { DOKUMENTARKIV_URL } from "astro:env/server";
-import { BodyShort } from "@navikt/ds-react";
-import { formatDateMonth, setAvsenderMottaker } from "./dokumentUtils";
-import type { DokumentType } from "../dokumenterTypes";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
+import { BodyShort } from "@navikt/ds-react";
+import type { DokumentType } from "@src/features/dokumenter/dokumenterTypes";
 import styles from "./Dokument.module.css";
+import { formatDateMonth, setAvsenderMottaker } from "./dokumentUtils";
 
 interface Props {
   dokument: DokumentType;
@@ -30,7 +30,7 @@ const avsender = setAvsenderMottaker(dokument);
 </li>
 
 <script>
-  import { logEvent } from "../../../utils/client/umami";
+  import { logEvent } from "@src/shared/utils/client/umami";
 
   const dokumentlenke = document.querySelector("#dokumentlenke");
