+++ src/features/dokumenter/ingen/MinIDDokumenter.astro
@@ -1,8 +1,8 @@
 ---
-import { text } from "../dokumenterText";
 import { BodyLong, Heading } from "@navikt/ds-react";
-import { Language } from "../../../utils/server/language";
-import styles from "../Dokumenter.module.css";
+import styles from "@src/features/dokumenter/Dokumenter.module.css";
+import { text } from "@src/features/dokumenter/dokumenterText";
+import type { Language } from "@src/shared/utils/server/language";
 
 interface Props {
   language: Language;
