+++ src/features/innloggede-tjenester/InnloggedeTjenester.astro
@@ -1,6 +1,6 @@
 ---
 import { BodyShort, Heading } from "@navikt/ds-react";
-import type { Language } from "../../utils/server/language";
+import type { Language } from "@src/shared/utils/server/language";
 import styles from "./InnloggedeTjenester.module.css";
 import { text } from "./innloggedeTjenesterText";
 import { annetLenker, hjelpemidlerLenker, jobbLenker, personopplysningLenker } from "./innloggedeTjenesterUrls";
