+++ src/features/innloggede-tjenester/section/InnloggedeTjenesterSection.astro
@@ -1,8 +1,8 @@
 ---
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
 import { BodyShort, Detail } from "@navikt/ds-react";
-import type { Language } from "../../../utils/server/language";
-import LinkWithAmplitude from "../link/LinkWithAmplitude";
+import LinkWithAmplitude from "@src/features/innloggede-tjenester/link/LinkWithAmplitude";
+import type { Language } from "@src/shared/utils/server/language";
 import style from "./InnloggedeTjensterSection.module.css";
 
 const language = Astro.currentLocale as Language;
