+++ src/features/personalia/Navn.astro
@@ -1,9 +1,9 @@
 ---
-import { getAudience, getOboToken } from "../../utils/server/token";
-import ClientError from "@shared/client-error/ClientError";
 import { NAVN_URL } from "astro:env/server";
-import { fetchData } from "../../utils/server/fetch";
-import logger from "../../utils/server/logger";
+import ClientError from "@src/shared/client-error/ClientError";
+import { fetchData } from "@src/shared/utils/server/fetch";
+import logger from "@src/shared/utils/server/logger";
+import { getAudience, getOboToken } from "@src/shared/utils/server/token";
 import style from "./Personalia.module.css";
 
 const audience = getAudience("tms-min-side-proxy");
