+++ src/features/personalia/Personalia.astro
@@ -1,5 +1,5 @@
 ---
-import type { Language } from "../../utils/server/language";
+import type { Language } from "@src/shared/utils/server/language";
 import Navn from "./Navn.astro";
 import style from "./Personalia.module.css";
 import { text } from "./personaliaText";
