+++ src/features/substantial-info/SubstantialInfo.astro
@@ -1,6 +1,6 @@
 ---
 import { Alert, BodyShort, Heading, Link } from "@navikt/ds-react";
-import type { Language } from "../../utils/server/language";
+import type { Language } from "@src/shared/utils/server/language";
 import { text } from "./substantialInfoText";
 
 const language = Astro.currentLocale as Language;
