+++ src/features/utbetaling/Utbetaling.astro
@@ -1,11 +1,11 @@
 ---
 import { UTBETALINGSOVERSIKT_API_URL } from "astro:env/server";
-import ClientError from "@shared/client-error/ClientError";
-import PersonalContentLogger from "@shared/legacy/PersonalContentLogger";
-import logger from "@utils/server/logger";
-import { fetchData } from "../../utils/server/fetch";
-import type { Language } from "../../utils/server/language";
-import { getAudience, getOboToken } from "../../utils/server/token";
+import ClientError from "@src/shared/client-error/ClientError";
+import PersonalContentLogger from "@src/shared/legacy/PersonalContentLogger";
+import { fetchData } from "@src/shared/utils/server/fetch";
+import type { Language } from "@src/shared/utils/server/language";
+import logger from "@src/shared/utils/server/logger";
+import { getAudience, getOboToken } from "@src/shared/utils/server/token";
 import EnkelUtbetaling from "./enkel/EnkelUtbetaling.astro";
 import IngenUtbetaling from "./ingen/IngenUtbetaling.astro";
 import UtbetalingList from "./list/UtbetalingList.astro";
