+++ src/features/utbetaling/enkel/EnkelUtbetaling.astro
@@ -1,12 +1,12 @@
 ---
-import { Heading, BodyShort } from "@navikt/ds-react";
-import { text } from "../utbetalingText";
-import { formatToReadableDate } from "../utbetalingUtils";
-import type { UtbetalingData } from "../utbetalingTypes";
-import type { Language } from "../../../utils/server/language";
-import styles from "./EnkelUtbetaling.module.css";
 import { UTBETALINGSOVERSIKT_URL } from "astro:env/server";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
+import { BodyShort, Heading } from "@navikt/ds-react";
+import { text } from "@src/features/utbetaling/utbetalingText";
+import type { UtbetalingData } from "@src/features/utbetaling/utbetalingTypes";
+import { formatToReadableDate } from "@src/features/utbetaling/utbetalingUtils";
+import type { Language } from "@src/shared/utils/server/language";
+import styles from "./EnkelUtbetaling.module.css";
 
 interface Props {
   utbetaling: UtbetalingData;
@@ -47,7 +47,7 @@ const seDetaljerStyling = {
   >
 </a>
 <script>
-  import { logEvent } from "../../../utils/client/umami";
+  import { logEvent } from "@src/shared/utils/client/umami";
 
   const kommende = document.querySelector("#kommende");
   const tidligere = document.querySelector("#tidligere");
