+++ src/features/utbetaling/fallback/UtbetalingFallback.astro
@@ -1,8 +1,8 @@
 ---
 import { UTBETALINGSOVERSIKT_URL } from "astro:env/server";
 import { Heading, Skeleton } from "@navikt/ds-react";
-import type { Language } from "../../../utils/server/language";
-import { text } from "../utbetalingText";
+import { text } from "@src/features/utbetaling/utbetalingText";
+import type { Language } from "@src/shared/utils/server/language";
 import style from "./UtbetalingFallback.module.css";
 
 const language = Astro.currentLocale as Language;
