+++ src/features/utbetaling/ingen/IngenUtbetaling.astro
@@ -1,7 +1,7 @@
 ---
-import { Heading, BodyShort } from "@navikt/ds-react";
-import { Language } from "../../../utils/server/language";
-import { text } from "../utbetalingText";
+import { BodyShort, Heading } from "@navikt/ds-react";
+import { text } from "@src/features/utbetaling/utbetalingText";
+import type { Language } from "@src/shared/utils/server/language";
 import styles from "./IngenUtbetaling.module.css";
 
 interface Props {
