+++ src/features/utbetaling/se-alle/SeAlle.astro
@@ -1,9 +1,9 @@
 ---
 import { UTBETALINGSOVERSIKT_URL } from "astro:env/server";
-import { text } from "../utbetalingText";
-import { Language } from "../../../utils/server/language";
-import styles from "./SeAlle.module.css";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
+import { text } from "@src/features/utbetaling/utbetalingText";
+import type { Language } from "@src/shared/utils/server/language";
+import styles from "./SeAlle.module.css";
 
 interface Props {
   language: Language;
@@ -23,7 +23,7 @@ const { language, isKommende } = Astro.props;
     <Chevron className={styles.chevron} fontSize="24px" aria-hidden={true} />
   </a>
   <script>
-    import { logEvent } from "../../../utils/client/umami";
+    import { logEvent } from "@src/shared/utils/client/umami";
 
     const seAlle = document.querySelector("#seAlle");
