+++ src/microfrontends/Microfrontend.astro
@@ -1,10 +1,10 @@
 ---
-import { getAudience, getOboToken } from "../utils/server/token";
-import { fetchHtml } from "../utils/server/fetch";
-import { Microfrontend } from "./microfrontendTypes";
-import ClientError from "@shared/client-error/ClientError";
-import logger from "../utils/server/logger";
+import ClientError from "@src/shared/client-error/ClientError";
+import { fetchHtml } from "@src/shared/utils/server/fetch";
+import logger from "@src/shared/utils/server/logger";
+import { getAudience, getOboToken } from "@src/shared/utils/server/token";
 import styles from "./Microfrontend.module.css";
+import type { Microfrontend } from "./microfrontendTypes";
 
 interface Props {
   microfrontend: Microfrontend;
