+++ src/pages/[locale]/404.astro
@@ -1,6 +1,6 @@
 ---
-import Layout from "../../layouts/base/Layout.astro";
-import Content from "../../layouts/content/Content.astro";
+import Layout from "@src/layouts/base/Layout.astro";
+import Content from "@src/layouts/content/Content.astro";
 ---
 
 <Layout title="Min side">
