+++ src/pages/[locale]/index.astro
@@ -1,24 +1,24 @@
 ---
 import { PUBLIC_LEGACY_SELECTOR_URL, PUBLIC_TELEMETRY_URL } from "astro:env/server";
-import AlertIsland from "@features/alert-island/AlertIsland.astro";
-import Innboks from "@features/innboks/Innboks.astro";
-import InnloggedeTjenester from "@features/innloggede-tjenester/InnloggedeTjenester.astro";
-import Personalia from "@features/personalia/Personalia.astro";
-import SubstantialInfo from "@features/substantial-info/SubstantialInfo.astro";
-import Utbetaling from "@features/utbetaling/Utbetaling.astro";
-import Container from "@shared/container/Container.astro";
-import Feilmelding from "@shared/feilmelding/Feilmelding";
-import Legacy from "@shared/legacy/Legacy";
-import Observability from "@shared/obersvability/Observability";
-import Aktuelt from "../../features/aktuelt/Aktuelt.astro";
-import DinOversikt from "../../features/din-oversikt/DinOversikt.astro";
-import Dokumenter from "../../features/dokumenter/Dokumenter.astro";
-import DokumenterFallback from "../../features/dokumenter/fallback/DokumenterFallback.astro";
-import InnboksFallback from "../../features/innboks/fallback/InnboksFallback.astro";
-import UtbetalingFallback from "../../features/utbetaling/fallback/UtbetalingFallback.astro";
-import Layout from "../../layouts/base/Layout.astro";
-import Content from "../../layouts/content/Content.astro";
-import type { Language } from "../../utils/server/language";
+import Aktuelt from "@src/features/aktuelt/Aktuelt.astro";
+import AlertIsland from "@src/features/alert-island/AlertIsland.astro";
+import DinOversikt from "@src/features/din-oversikt/DinOversikt.astro";
+import Dokumenter from "@src/features/dokumenter/Dokumenter.astro";
+import DokumenterFallback from "@src/features/dokumenter/fallback/DokumenterFallback.astro";
+import InnboksFallback from "@src/features/innboks/fallback/InnboksFallback.astro";
+import Innboks from "@src/features/innboks/Innboks.astro";
+import InnloggedeTjenester from "@src/features/innloggede-tjenester/InnloggedeTjenester.astro";
+import Personalia from "@src/features/personalia/Personalia.astro";
+import SubstantialInfo from "@src/features/substantial-info/SubstantialInfo.astro";
+import UtbetalingFallback from "@src/features/utbetaling/fallback/UtbetalingFallback.astro";
+import Utbetaling from "@src/features/utbetaling/Utbetaling.astro";
+import Layout from "@src/layouts/base/Layout.astro";
+import Content from "@src/layouts/content/Content.astro";
+import Container from "@src/shared/container/Container.astro";
+import Feilmelding from "@src/shared/feilmelding/Feilmelding";
+import Legacy from "@src/shared/legacy/Legacy";
+import Observability from "@src/shared/obersvability/Observability";
+import type { Language } from "@src/shared/utils/server/language";
 
 const language = Astro.currentLocale as Language;
 ---
