+++ src/shared/authentication/Authentication.tsx
@@ -1,6 +1,6 @@
+import { fetcher } from "@src/shared/utils/client/api.ts";
 import useSWRImmutable from "swr/immutable";
 import { redirectUrl, statusUrl } from "./authenticationUrls";
-import { fetcher } from "@utils/client/api.ts";
 
 export const reloadOnPageshow = () => {
   window.addEventListener("pageshow", (event) => {
