+++ src/shared/authentication/authenticationUrls.ts
@@ -1,4 +1,4 @@
-import { getDevBaseUrl, getEnvironmentClientSide } from "@utils/client/environment.ts";
+import { getDevBaseUrl, getEnvironmentClientSide } from "@src/shared/utils/client/environment.ts";
 
 const MINSIDE_URL = {
   local: "http://localhost:4321/minside",
