+++ src/shared/client-error/ClientError.tsx
@@ -1,4 +1,4 @@
-import { setIsError } from "../store/store";
+import { setIsError } from "@src/shared/store/store";
 
 const ClientError = () => {
   setIsError();
