+++ src/shared/legacy/legacyUtils.ts
@@ -1,7 +1,7 @@
-import { Language } from "../../utils/server/language.ts";
-import { Microfrontend } from "../../microfrontends/microfrontendTypes";
-import { PersonalizedContent } from "@features/din-oversikt/DinOversiktTypes";
-import { produktText } from "@features/din-oversikt/produktkort/ProduktText";
+import type { PersonalizedContent } from "@src/features/din-oversikt/DinOversiktTypes";
+import { produktText } from "@src/features/din-oversikt/produktkort/ProduktText";
+import type { Microfrontend } from "@src/microfrontends/microfrontendTypes";
+import type { Language } from "@src/shared/utils/server/language.ts";
 
 export const getProduktPropertiesLegacy = (language: Language, personalizedContent?: PersonalizedContent) => {
   if (personalizedContent === undefined) return undefined;
