+++ src/shared/utils/client/environment.ts
