+++ src/shared/utils/server/error.ts
